% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bars.R
\name{CheckNewBarParams}
\alias{CheckNewBarParams}
\title{Checking for new params}
\usage{
CheckNewBarParams(symbol = "", startTime = "", endTime = "",
  periodicity = "M1", priceType = "Bid")
}
\description{
Checking for new params
}
