% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{CheckNewTickParams}
\alias{CheckNewTickParams}
\title{Checking for new params}
\usage{
CheckNewTickParams(symbol = "", startTime = "", endTime = "",
  level2 = FALSE)
}
\description{
Checking for new params
}
