% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bars.R
\name{tthBarRequest}
\alias{tthBarRequest}
\title{Gets the bars as requested}
\usage{
tthBarRequest(symbol = "", timestamp = "", count = "",
  periodicity = "M1", priceType = "Bid")
}
\arguments{
\item{symbol}{Symbol looked}

\item{timestamp}{timestamp}

\item{count}{Count of bars}

\item{periodicity}{periodicity}

\item{priceType}{priceType}
}
\description{
Gets the bars as requested
}
