% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{tthConnect}
\alias{tthConnect}
\title{Connects to a TT server}
\usage{
tthConnect(address = "", login = "", password = "", port = 5020,
  name = paste0("client_", Sys.getpid()))
}
\arguments{
\item{address}{Url of the server}

\item{login}{account you login}

\item{password}{password of the account}

\item{port}{port}

\item{name}{name of client}
}
\description{
Connects to a TT server
}
