% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bars.R
\name{tthFileBarRequest}
\alias{tthFileBarRequest}
\title{Gets the bars as requested}
\usage{
tthFileBarRequest(symbol = "", startTime = "", endTime = "",
  periodicity = "M1", priceType = "Bid")
}
\arguments{
\item{symbol}{Symbol looked}

\item{startTime}{Start of the time intervals}

\item{endTime}{Count of bars}

\item{periodicity}{periodicity}

\item{priceType}{priceType}
}
\description{
Gets the bars as requested
}
