% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{tthFileTickRequest}
\alias{tthFileTickRequest}
\title{Gets the ticks as requested}
\usage{
tthFileTickRequest(symbol = "", startTime = "", endTime = "")
}
\arguments{
\item{symbol}{Symbol looked}

\item{startTime}{Start of the time intervals}

\item{endTime}{Count of ticks}
}
\description{
Gets the ticks as requested
}
