% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lists.R
\name{tthGetSupportedPeriodicities}
\alias{tthGetSupportedPeriodicities}
\title{Gets the list of supported periodicities}
\usage{
tthGetSupportedPeriodicities()
}
\description{
Gets the list of supported periodicities
}
