% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TicksL2.R
\name{tthTickL2Request}
\alias{tthTickL2Request}
\title{Gets the ticks level 2 as requested}
\usage{
tthTickL2Request(symbol = "", timestamp = "", count = "")
}
\arguments{
\item{symbol}{Symbol looked}

\item{timestamp}{timestamp}

\item{count}{Count of ticks}
}
\description{
Gets the ticks level 2 as requested
}
