% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bars.R
\name{tthBarRequest}
\alias{tthBarRequest}
\title{Gets the bars as requested}
\usage{
tthBarRequest(symbol = "", endTime = "", count = "", periodicity = "M1",
  priceType = "Bid")
}
\arguments{
\item{symbol}{Symbol looked}

\item{count}{Count of ticks}

\item{startTime}{Start of the time intervals}
}
\description{
Gets the bars as requested
}
