% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TicksL2.R
\name{tthFileTickL2Request}
\alias{tthFileTickL2Request}
\title{Gets the ticks level 2 as requested}
\usage{
tthFileTickL2Request(symbol = "", startTime = "", endTime = "",
  depth = 0)
}
\arguments{
\item{symbol}{Symbol looked}

\item{startTime}{Start of the time intervals}

\item{endTime}{Count of ticks}

\item{depth}{Depth of ticks}
}
\description{
Gets the ticks level 2 as requested
}
