% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TicksL2.R
\name{tthStreamTickL2Request}
\alias{tthStreamTickL2Request}
\title{Gets the ticks level 2 as requested}
\usage{
tthStreamTickL2Request(symbol = "", startTime = "", endTime = "")
}
\arguments{
\item{symbol}{Symbol looked}

\item{startTime}{Start of the time intervals}

\item{endTime}{Count of ticks}
}
\description{
Gets the ticks level 2 as requested
}
