% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{tthTickRequest}
\alias{tthTickRequest}
\title{Gets the ticks as requested}
\usage{
tthTickRequest(symbol = "", endTime = "", count = "")
}
\arguments{
\item{symbol}{Symbol looked}

\item{endTime}{Start of the time intervals}

\item{count}{Count of ticks}
}
\description{
Gets the ticks as requested
}
